svnexternals: preliminary support for subrepos based externals
At this point, only pulling externals definitions into .hgsub and .hgsubstate
is supported. One difference between subrepos and svn:externals is the former
separate the source definition and target revision in two files, while
svn:externals definitions contain both. To handle this, the svn:externals
revision references is replaced with a {REV} placeholder and stored in .hgsub,
prefixed with the external base directory separated with a ':', while the
revision is extracted in .hgsubstate.

For instance, the following external:

    -r3 ^/externals/proj2@2 deps/proj2

Becomes:

    (.hgsub)
    deps/proj2 = [hgsubversion] :-r{REV} ^/externals/proj2@2 deps/proj2

    (.hgsubstate)
    3 deps/proj2