stupid: diff with the good revision when restoring branch
Converting history like:

  r1 create trunk
  r2 remove trunk
  r3 make some unrelated commit
  r4 restore trunk from r2

stupid mode would retrieve and apply the diff with r3 instead of the
one with r2. All the code deciding r2 is the actual parent was already
there, only the revision argument was not passed.

This fixes:
- test_delete_restore_trunk in comprehensive/test_stupid_pull.py
- test_delete_restore_trunk_stupid in comprehensive/test_verify_and_startrev.py

Tested on OSX 10.6 with macports svn-1.7.5. I can no longer test with
svn-1.6.