editor: relax copyfrom dir checks to avoid extra missing entries
When renaming a branch you get something like:

  D /branch/bar
  A /branch/foo (from /branch/foo:42)

Unfortunately, the branch layout for the revision being converted is
computed before starting to convert it. It means the copyfrom path
supplied in the add_directory() for /branch/foo will be be considered
invalid, be added to missing and fetched the slow way despite being in
the repository history. Avoid that by checking the path looks like a
branch path and matching it with the filemap. It will be resolved
afterwards anyway.