bundle: add 'portable' parameter to prevent error (issue383)
hgsubversion does not fully implement the -r/--rev flag for `hg bundle` so
would crash when passed this parameter since the wrapped findcommonoutgoing
did not match the parameter list hg expects.  With this patch the portable
parameter is added to this function, but remains unused so that `hg bundle -r`
will properly generate a bundle, but up to tip (in essence ignoring -r
parameter).