util: add visitor pattern for scrubbing json
These functions are for future patches that will add safer serialization via
json. '_convert' is a visitor pattern that will be used for lists,
dictionaries, and strings for helping convert None to the empty string since
json forbids 'null' as a key for a dictionary.

None -> '' is a safe mapping because this is for the 'branch_info' variable
which already maps the empty string to None.

Note, also, that json is chosen instead of, say, csv because json has a concept
of 'null' and will better handle utf8 strings (which subversion supports).

Important: this changes the requirement of hgsubversion to python 2.6+.