svnmeta: add generator for cached svn configuration properties
This will allow use to unify the reading and writing of configuration options
into a central object to simplify their use sprinkled throughout the codebase.

The idea is that after this patch, we will move each option to the svnmeta
class thereby allowing us to remove lots of I/O cruft. Once the cruft is gone,
we'll refactor objects where necessary. After refactoring, we'll have a
framework for easily adding new configuration options.