editor: correctly import symlink copy+modify with non-empty prefix
We alwasy fail editing for symlinks, since we strip the leading 'link '
subversion includes when storing in mercurial, and then let svn
attempt to apply deltas against the stripped version.  This
unsurprisingly fails, and we write the resulting empty-string to the
Filestore for the current revision, and add the symlink in question to
the missing list to handle stupidly later.

Unfortunately, this would break down because editing adds files to the
store using their absolute path whereas missing files are added
relative to our subdir.  the absolut path file appears to win, which
results in us getting a symlink whose target is the empty string.

This fixes the problem by adding missing files to the fileStore using
their absolute path.