tests: use hostname instead of localhost
The test_push_command test was failing on our machines because the machines are
ipv6 but expose ipv4 on the 127.0.0.1 loopback interface. This caused the
svnserve process to listen via ipv4, but the connecting process would attempt to
connect on ipv6 and fail. Using the hostname causes it to listen using the
primary network interface, which matches the interface that is used when the
connecting client resolves the hostname. So the test now passes.